Moved level editor into core module and incorporated it into base game
