+++ core/src/main/java/core/gdx/wad/StartMenu.java
@@ -1,9 +1,7 @@
 package core.gdx.wad;
 
 import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
@@ -11,14 +9,8 @@ import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
-import com.badlogic.gdx.utils.Align;
-import core.game.logic.GameLogic;
-import core.level.info.LevelData;
-import core.server.Network;
 import core.server.SpaceClient;
-import org.lwjgl.system.CallbackI;
-
-import java.util.function.BiConsumer;
+import editor.launch.EditorScreen;
 
 public class StartMenu extends Actor{
 
@@ -191,6 +183,16 @@ public class StartMenu extends Actor{
                 super.clicked(event, x, y);
                 System.out.println("Settings\n");
                 myGDxTest.setScreen(settingsScreen);
+                //((MyGDxTest) Gdx.app.getApplicationListener()).setScreen(new SettingsScreen());
+            }
+        });
+
+        levelEditorButton.addListener(new ClickListener() {
+            @Override
+            public void clicked(InputEvent event, float x, float y) {
+                super.clicked(event, x, y);
+                System.out.println("Editor\n");
+                myGDxTest.setScreen(new EditorScreen());
                 //((MyGDxTest) Gdx.app.getApplicationListener()).setScreen(new SettingsScreen());
             }
         });
