+++ core/src/main/java/editor/windows/LevelChooserWindow.java
@@ -1,4 +1,4 @@
-package editor.gdx.windows;
+package editor.windows;
 
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
@@ -6,8 +6,8 @@ import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.utils.Array;
 import core.level.info.LevelData;
 import core.wad.funcs.WadFuncs;
-import editor.gdx.launch.EditorScreen;
-import editor.gdx.windows.actors.NumberField;
+import editor.launch.EditorScreen;
+import editor.windows.actors.NumberField;
 import net.mtrop.doom.WadEntry;
 import net.mtrop.doom.WadFile;
