+++ core/src/main/java/editor/windows/actors/DigitsPlusNegativeFilter.java
@@ -1,4 +1,4 @@
-package editor.gdx.windows.actors;
+package editor.windows.actors;
 
 import com.badlogic.gdx.scenes.scene2d.ui.TextField;
