+++ core/src/main/java/editor/windows/actors/NumberField.java
@@ -1,11 +1,11 @@
-package editor.gdx.windows.actors;
+package editor.windows.actors;
 
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.TextField;
 
 public class NumberField extends TextField {
 
-    final private static TextField.TextFieldFilter numberFilter = new DigitsPlusNegativeFilter();
+    final private static TextFieldFilter numberFilter = new DigitsPlusNegativeFilter();
 
     public NumberField(String text, Skin skin) {
         super(text, skin);
