--- editor/build.gradle
@@ -1,9 +1,0 @@
-[compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
-eclipse.project.name = appName + '-core'
-
-dependencies {
-	implementation project(':core')
-	api "com.badlogicgames.gdx:gdx:$gdxVersion"
-	api "com.badlogicgames.gdx:gdx-box2d:$gdxVersion"
-	api files('../libs/doomstruct-2.14.0.jar')
-}
