--- editor/src/main/java/editor/gdx/launch/LevelEditor.java
@@ -1,20 +1,0 @@
-package editor.gdx.launch;
-
-import com.badlogic.gdx.Game;
-import core.game.logic.GameLogic;
-import core.wad.funcs.WadFuncs;
-
-/** {@link com.badlogic.gdx.ApplicationListener} implementation shared by all platforms. */
-public class LevelEditor extends Game {
-    public EditorScreen editorScreen;
-
-    public LevelEditor() {
-        super();
-    }
-
-    @Override
-    public void create() {
-        editorScreen = new EditorScreen(this);
-        setScreen(editorScreen);
-    }
-}
