--- lwjgl3/src/main/java/core/gdx/wad/lwjgl3/EditorLauncher.java
@@ -1,24 +1,0 @@
-package core.gdx.wad.lwjgl3;
-
-import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
-import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
-import editor.gdx.launch.LevelEditor;
-
-public class EditorLauncher {
-
-    public static void main(String[] args) {
-        createApplication();
-    }
-
-    private static Lwjgl3Application createApplication() {
-        return new Lwjgl3Application(new LevelEditor(), getDefaultConfiguration());
-    }
-
-    private static Lwjgl3ApplicationConfiguration getDefaultConfiguration() {
-        Lwjgl3ApplicationConfiguration configuration = new Lwjgl3ApplicationConfiguration();
-        configuration.setTitle("Space Crucible Level Editor");
-        configuration.setWindowedMode(640, 480);
-        configuration.setWindowIcon("libgdx128.png", "libgdx64.png", "libgdx32.png", "libgdx16.png");
-        return configuration;
-    }
-}
