--- lwjgl3/src/main/java/core/gdx/wad/lwjgl3/LegacyEditorLauncher.java
@@ -1,24 +1,0 @@
-package core.gdx.wad.lwjgl3;
-
-import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
-import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import editor.gdx.launch.LevelEditor;
-
-/** Launches the desktop (LWJGL) application. */
-public class LegacyEditorLauncher {
-	public static void main(String[] args) {
-		createApplication();
-	}
-
-	private static LwjglApplication createApplication()  {
-		return new LwjglApplication(new LevelEditor(), getDefaultConfiguration());
-	}
-
-	private static LwjglApplicationConfiguration getDefaultConfiguration() {
-		LwjglApplicationConfiguration configuration = new LwjglApplicationConfiguration();
-		configuration.title = "Space Crucible";
-		configuration.width = 640;
-		configuration.height = 480;
-		return configuration;
-	}
-}
